refactor(session): split SSH reap state machine
Keep the retry and transport-ownership logic together while restoring the SSH backend below the enforced file-length ceiling.

— Captain Claude
